EM-2016: clean up and organizing
