+++ src/components/vanilla/controls/MultiSelectorDimensionOrMeasure/index.tsx
@@ -1,31 +1,29 @@
 import { DimensionOrMeasure } from '@embeddable.com/core';
 import React from 'react';
 
-import MultiSelector from '../../MultiSelector';
+import MultiSelector from '../MultiSelector';
+import { getSelectorOptions } from '../Selector.utils';
 
 export type Props = {
   allowNoValue?: boolean;
-  defaultValue: DimensionOrMeasure[];
+  defaultValue?: DimensionOrMeasure[];
   options: DimensionOrMeasure[];
   title?: string;
   onChange: (v: DimensionOrMeasure[]) => void;
 };
 
 export default (props: Props) => {
-  const calculatedDefaultValue = (props.defaultValue || []).map((v) => v.name);
+  const calculatedDefaultValue = props.defaultValue?.map((v) => v.name) || [];
 
   const handleChange = (newValues: string[]) => {
-    const newSelection = props.options.filter((option) => newValues.includes(option.name))!;
+    const newSelection = props.options.filter((option) => newValues.includes(option.name));
     props.onChange(newSelection);
   };
 
   return (
     <MultiSelector
       title={props.title}
-      options={props.options.map((option) => ({
-        value: option.name,
-        label: option.inputs?.overrideName ?? option.title,
-      }))}
+      options={getSelectorOptions(props.options)}
       defaultValue={calculatedDefaultValue}
       unclearable={!props.allowNoValue}
       onChange={handleChange}
