+++ src/components/vanilla/controls/MultiSelectorMeasure/index.tsx
@@ -1,31 +1,29 @@
-import { Dimension, Measure } from '@embeddable.com/core';
+import { Measure } from '@embeddable.com/core';
 import React from 'react';
 
-import MultiSelector from '../../MultiSelector';
+import MultiSelector from '../MultiSelector';
+import { getSelectorOptions } from '../Selector.utils';
 
 export type Props = {
   allowNoValue?: boolean;
-  defaultValue: Measure[];
+  defaultValue?: Measure[];
   options: Measure[];
   title?: string;
   onChange: (v: Measure[]) => void;
 };
 
 export default (props: Props) => {
-  const calculatedDefaultValue = (props.defaultValue || []).map((v) => v.name);
+  const calculatedDefaultValue = props.defaultValue?.map((v) => v.name) || [];
 
   const handleChange = (newValues: string[]) => {
-    const newSelection = props.options.filter((option) => newValues.includes(option.name))!;
+    const newSelection = props.options.filter((option) => newValues.includes(option.name));
     props.onChange(newSelection);
   };
 
   return (
     <MultiSelector
       title={props.title}
-      options={props.options.map((option) => ({
-        value: option.name,
-        label: option.inputs?.overrideName ?? option.title,
-      }))}
+      options={getSelectorOptions(props.options)}
       defaultValue={calculatedDefaultValue}
       unclearable={!props.allowNoValue}
       onChange={handleChange}
