+++ src/components/vanilla/controls/Selector.tsx
@@ -9,10 +9,8 @@ import React, {
   useState,
 } from 'react';
 import { twMerge } from 'tailwind-merge';
-import Checkbox from '../icons/Checkbox';
-import CheckboxEmpty from '../icons/CheckboxEmpty';
-import Container from './Container';
-import { ChevronDown, ClearIcon } from './icons';
+import Container from '../Container';
+import { ChevronDown, ClearIcon } from '../icons';
 import { SelectorOption, SelectorRecord } from './Selector.types';
 
 export type Props = {
@@ -20,11 +18,11 @@ export type Props = {
   searchProperty?: string;
   minDropdownWidth?: number;
   placeholder?: string;
-  defaultValue?: string[];
+  defaultValue?: string;
   options: SelectorOption[];
   title?: string;
   unclearable?: boolean;
-  onChange: (v: string[]) => void;
+  onChange: (v: string) => void;
 };
 
 let debounce: number | undefined = undefined;
@@ -37,6 +35,10 @@ export default (props: Props) => {
   const [search, setSearch] = useState('');
   const [triggerBlur, setTriggerBlur] = useState(false);
   const [value, setValue] = useState(props.defaultValue);
+
+  const valueLabel: string | undefined = props.options.find(
+    (option) => option.value === value,
+  )?.label;
 
   const [, setServerSearch] = useEmbeddableState({
     [props.searchProperty || 'search']: '',
@@ -91,28 +93,13 @@ export default (props: Props) => {
     [setSearch, setServerSearch, props.searchProperty],
   );
 
-  const set = useCallback(
-    (newValue: string) => {
+  const setDropdownValue = useCallback(
+    (value: string) => {
       performSearch('');
-
-      let newValues: string[] = [];
-
-      if (newValue !== '') {
-        newValues = value || [];
-        if (newValues?.includes(newValue)) {
-          newValues = newValues.filter((v) => v !== newValue);
-        } else {
-          newValues = [...newValues, newValue];
-        }
-      }
-
-      props.onChange(newValues);
-      setValue(newValues);
-      setServerSearch((s) => ({ ...s, [props.searchProperty || 'search']: '' }));
-      clearTimeout(debounce);
+      setValue(value);
+      props.onChange(value);
     },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [performSearch, props, value],
+    [setValue, props, performSearch],
   );
 
   // Used for handling keydown events on the menu items
@@ -124,6 +111,8 @@ export default (props: Props) => {
     if (e.key === 'Enter' || e.key === ' ') {
       e.preventDefault();
       callback(e);
+      setFocus(false);
+      setTriggerBlur(true);
     }
     if (escapable && e.key === 'Escape') {
       e.preventDefault();
@@ -133,36 +122,33 @@ export default (props: Props) => {
   };
 
   const list = useMemo(() => {
-    return props.options.map((option) => {
-      return (
-        <div
-          key={option.value}
-          role="button"
-          onClick={() => {
-            setTriggerBlur(false);
-            set(option.value);
-          }}
-          onKeyDown={(e) => handleKeyDownCallback(e, set(option.value), true)}
-          onFocus={() => {
-            setIsDropdownOrItemFocused(true);
-            setFocus(true);
-          }}
-          onBlur={() => {
-            setIsDropdownOrItemFocused(false);
-          }}
-          className={`flex items-left items-center min-h-[36px] px-3 py-2 hover:bg-black/5 cursor-pointer font-normal ${
-            value?.includes(option.value) ? 'bg-black/5' : ''
-          } truncate`}
-          tabIndex={0}
-        >
-          {value?.includes(option.value) ? <Checkbox /> : <CheckboxEmpty />}
-          <span className="font-normal pl-1 truncate" title={option.label}>
-            {option.label}
-          </span>
-        </div>
-      );
-    });
-  }, [props, value, set]) as ReactNode[];
+    return props.options?.map((option) => (
+      <div
+        key={option.value}
+        role="button"
+        onClick={() => {
+          setFocus(false);
+          setTriggerBlur(true);
+          setDropdownValue(option.value);
+        }}
+        onKeyDown={(e) => {
+          handleKeyDownCallback(
+            e,
+            () => {
+              setDropdownValue(option.value);
+            },
+            true,
+          );
+        }}
+        className={`flex items-center min-h-[36px] px-3 py-2 hover:bg-black/5 cursor-pointer font-normal ${
+          value === option.value ? 'bg-black/5' : ''
+        } whitespace-nowrap overflow-hidden text-ellipsis`}
+        tabIndex={0}
+      >
+        {option.label}
+      </div>
+    ));
+  }, [props.options, value, setDropdownValue]) as ReactNode[];
 
   return (
     <Container title={props.title}>
@@ -195,13 +181,13 @@ export default (props: Props) => {
           }`}
         />
 
-        {!!value && (
+        {valueLabel && (
           <span
-            className={`absolute w-[calc(100%-2rem)] whitespace-nowrap overflow-hidden truncate rounded-xl left-3 top-1 h-8 leading-8 block pointer-events-none text-sm ${
+            className={`absolute w-[calc(100%-2.5rem)] whitespace-nowrap overflow-hidden truncate rounded-xl left-3 top-1 h-8 leading-8 block pointer-events-none text-sm ${
               focus ? 'hidden' : ''
             }`}
           >
-            Selected {value.length} {value.length === 1 ? 'option' : 'options'}
+            {valueLabel}
           </span>
         )}
 
@@ -230,12 +216,12 @@ export default (props: Props) => {
           </div>
         )}
 
-        <ChevronDown className="absolute right-2.5 top-2.5 z-1 pointer-events-none" />
+        <ChevronDown className="absolute right-2 top-2.5 z-1 pointer-events-none" />
 
         {!props.unclearable && !!value && (
           <div
             onClick={() => {
-              set('');
+              setDropdownValue('');
             }}
             className="absolute right-10 top-0 h-10 flex items-center z-10 cursor-pointer"
           >
