+++ src/components/vanilla/controls/SelectorMeasure/index.tsx
@@ -1,11 +1,12 @@
 import { Measure } from '@embeddable.com/core';
 import React from 'react';
 
-import Selector from '../../Selector';
+import Selector from '../Selector';
+import { getSelectorOptions } from '../Selector.utils';
 
 export type Props = {
   allowNoValue?: boolean;
-  defaultValue: Measure;
+  defaultValue?: Measure;
   options: Measure[];
   title?: string;
   onChange: (v: Measure) => void;
@@ -20,10 +21,7 @@ export default (props: Props) => {
   return (
     <Selector
       title={props.title}
-      options={props.options.map((option) => ({
-        value: option.name,
-        label: option.inputs?.overrideName ?? option.title,
-      }))}
+      options={getSelectorOptions(props.options)}
       defaultValue={props.defaultValue?.name}
       unclearable={!props.allowNoValue}
       onChange={handleChange}
